Explicit box/unboxing (compiler warning found some bugs)
git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1698313 13f79535-47bb-0310-9956-ffa450edef68

+++ src/components/org/apache/jmeter/assertions/SMIMEAssertion.java
@@ -335,8 +335,8 @@ class SMIMEAssertion {
             X509ExtensionUtil.getSubjectAlternativeNames(cert);
         for (Object altNameObj : subjectAltNames) {
             List<?> altName = (List<?>) altNameObj;
-            int type = (Integer) altName.get(0);
-            if (type == GeneralName.rfc822Name) {
+            Integer type = (Integer) altName.get(0);
+            if (type.intValue() == GeneralName.rfc822Name) {
                 String address = (String) altName.get(1);
                 res.add(address);
             }
